TST: Split test_tiatoolbox.py in multiple files - Split test_tiatoolbox.py in multiple files to clean the tests to avoid errors.
