+++ tests/test_utils.py
@@ -1,0 +1,22 @@
+from tiatoolbox import utils
+
+import numpy as np
+
+
+def test_imresize():
+    """Test for imresize."""
+    img = np.zeros((2000, 2000, 3))
+    resized_img = utils.transforms.imresize(img, 0.5)
+    assert resized_img.shape == (1000, 1000, 3)
+
+
+def test_background_composite():
+    """Test for background composite."""
+    new_im = np.zeros((2000, 2000, 4)).astype("uint8")
+    new_im[:1000, :, 3] = 255
+    im = utils.transforms.background_composite(new_im)
+    assert np.all(im[1000:, :, :] == 255)
+    assert np.all(im[:1000, :, :] == 0)
+
+    im = utils.transforms.background_composite(new_im, alpha=True)
+    assert np.all(im[:, :, 3] == 255)
